WC-454: client: Add timeout parameters.
This possibly breaks the async client, I couldn't see where the magical
class was that mapped its _send and _recv methods to corresponding
TwistedMatrix calls.

Essentially, when executing a transaction, there is now a timeout
parameter.  This is used to compute a deadline (based on `time.time`)
for the transaction.

The _send and _recv methods now also take a timeout, and this is used
with `socket.settimeout` on TCP/UDP and with the `timeout` and
`write_timeout` properties on Serial.

Before doing a read or write, if a deadline was calculated, this is used
to compute the remaining timeout for that read or write.  That enables
the whole operation to respect the timeout given.