WH-408: bit_read_message: Implement lazy decoding/encoding of bits.
Much like what we did for reading registers, this allows us to process
the decoding of the result when it is actually needed instead of doing
it all up-front.

Thus we remain backward compatible with other software that uses
`pymodbus` (including our own earlier code) and we gain a performance
boost when we want to hand-off decoding to other code (e.g.
`pyregblock`).